Invarinat mass histos added. Task isolated.
